+++ src/pyjion/absvalue.h
@@ -463,6 +463,7 @@ class FloatValue : public AbstractValue {
     AbstractValue* binary(AbstractSource* selfSources, int op, AbstractValueWithSources& other) override;
     AbstractValue* unary(AbstractSource* selfSources, int op) override;
     const char* describe() override;
+    AbstractValueKind resolveMethod(const char* name) override;
 public:
     static AbstractValue* binary(int op, AbstractValueWithSources& other);
 };